Move utils file one level down
